+++ packages/core-js/features/array/index.js
@@ -1,8 +1,8 @@
 var parent = require('../../es/array');
 require('../../modules/es.map');
+require('../../modules/esnext.array.at');
 require('../../modules/esnext.array.filter-out');
 require('../../modules/esnext.array.is-template-object');
-require('../../modules/esnext.array.item');
 require('../../modules/esnext.array.last-item');
 require('../../modules/esnext.array.last-index');
 require('../../modules/esnext.array.unique-by');