+++ packages/core-js/features/array/item.js
@@ -1,4 +1,0 @@
-require('../../modules/esnext.array.item');
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Array', 'item');