+++ packages/core-js/features/array/virtual/index.js
@@ -1,5 +1,6 @@
 var parent = require('../../../es/array/virtual');
+require('../../../modules/esnext.array.at');
 require('../../../modules/esnext.array.filter-out');
-require('../../../modules/esnext.array.item');
+require('../../../modules/esnext.array.unique-by');
 
 module.exports = parent;