+++ packages/core-js/features/string/item.js
@@ -1,4 +1,0 @@
-require('../../modules/esnext.string.item');
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('String', 'item');