+++ packages/core-js/features/string/virtual/item.js
@@ -1,4 +1,0 @@
-require('../../../modules/esnext.string.item');
-var entryVirtual = require('../../../internals/entry-virtual');
-
-module.exports = entryVirtual('String').item;