+++ packages/core-js/features/typed-array/index.js
@@ -1,5 +1,5 @@
 var parent = require('../../es/typed-array');
+require('../../modules/esnext.typed-array.at');
 require('../../modules/esnext.typed-array.filter-out');
-require('../../modules/esnext.typed-array.item');
 
 module.exports = parent;