+++ packages/core-js/proposals/relative-indexing-method.js
@@ -1,4 +1,5 @@
 // https://github.com/tc39/proposal-relative-indexing-method
-require('../modules/esnext.array.item');
-require('../modules/esnext.string.item');
-require('../modules/esnext.typed-array.item');
+require('../modules/esnext.array.at');
+// TODO: disabled by default because of the conflict with another proposal
+// require('../modules/esnext.string.at-alternative');
+require('../modules/esnext.typed-array.at');