+++ tests/compat/tests.js
@@ -1100,12 +1100,12 @@ GLOBAL.tests = {
       && set.add({}) == set
       && set[Symbol.toStringTag];
   }],
+  'esnext.array.at': function () {
+    return [].at;
+  },
   'esnext.array.filter-out': function () {
     return [].filterOut;
   }
-  'esnext.array.item': function () {
-    return [].item;
-  },
   'esnext.array.last-index': function () {
     return [1, 2, 3].lastIndex && Array.prototype[Symbol.unscopables].lastIndex;
   },
@@ -1430,9 +1430,6 @@ GLOBAL.tests = {
   'esnext.string.code-points': function () {
     return String.prototype.codePoints;
   },
-  'esnext.string.item': function () {
-    return ''.item;
-  },
   'esnext.symbol.dispose': function () {
     return Symbol.dispose;
   },
@@ -1446,12 +1443,12 @@ GLOBAL.tests = {
   'esnext.symbol.replace-all': function () {
     return Symbol.replaceAll;
   },
+  'esnext.typed-array.at': function () {
+    return Int8Array.prototype.at;
+  },
   'esnext.typed-array.filter-out': function () {
     return Int8Array.prototype.filterOut;
   }
-  'esnext.typed-array.item': function () {
-    return Int8Array.prototype.item;
-  },
   'esnext.weak-map.delete-all': function () {
     return WeakMap.prototype.deleteAll;
   },